warning C4267: conversion from 'size_t' to 'int', possible loss of data

+++ lclib/common/lc_mainwindow.cpp
@@ -2533,7 +2533,7 @@ void lcMainWindow::UpdateSelectedObjects(bool SelectionChanged, int SelectionTyp
 			{
 				QVector<TypeLine> LineTypeIndexes;
 
-				int Objects = Selection.size();
+				size_t Objects = Selection.size();
 				if (Objects > 0)
 				{
 					QString Message;
